Extract logger supplier tests
